requirements.txt
streamlit==1.36.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
scikit-image==0.24.0
albumentations==1.4.4
pillow==10.3.0
torch==2.2.0
torchvision==0.17.0
timm==0.9.16
